+++ frontend/jslib/xdialog/Changelog.md
@@ -1,6 +1,7 @@
 * v0.0.3 2012-10-25T15:59:48 UTC+0800
   * `read-only` 窗口调整
   * `Welcome` 窗口弹出 bug
+  * [+] `OAuth ForgotPassword` 窗口流程
 
 * v0.0.2 17:12:37 09/07/2012
   * `destory` 调整